Dialogs (#166)
* Comment out all calls to fix_dialog_size()

This is the first step in making the dialogs resizable in a
meaningful manner. This change allows them to resize but now we need
to make sure there are scrollbars and a widget manager in each dialog
so they can resize and still be useful.

* Add scrollbars to many dialogs

Xastir doesn't play well on low-resolution screens. This change will
allow users to shrink many dialogs, gaining scrollbars so the dialogs
can still be manipulated on smaller screens.

* Correct capitalization on label

Capitalization was wrong compared to other nearby labels.

* Correct length of Comment field for Object dialogs

Comment data field was overwriting widgets to the right. It is now
set to be smaller for display but the underlying data field can still
hold the max value it could before.

* Fix MAGICK ifdef's I messed up in the last 4 commits

I got out of sync somewhere with the master branch and committed
some unintended changes. This should fix that problem.

* Apply changes already in master plus a few more

Because my branch was messed up compared to master and would have removed
multiple commits already done if it were merged as it was. While I was at
it I got rid of some commented-out code, removed the fix_dialog_vsize()
function, and fixed another few spelling errors.