Fix handling of missing gust data in Davis/LaCrosse
I finally got my 2317 put back up after some home repair, and once
again was puzzled why the "gust" data in Xastir's "Own Weather Data"
was bouncing between a useful number and "...".

Turns out that the LaCrosse devices don't compute and report wind
gusts at all, and the LaCrosse "open2300db2APRS" program (from the
LaCrosse directory) simply sends "g..." for gust data unconditionally.

Then, in Xastir's code, the Davis Meteo section (which is what
the LaCrosse and wxnowsrv.pl tools masquerade as) unconditionally
reads the "gxxx" section of the report it gets and stuffs it into the
wx_gust piece of the weather struct.  Further, that code
unconditionally CLEARS the wx_gust data before it starts parsing the
incoming record.

But if we get g..., it means no valid data is coming from the weather
station, so we shouldn't try to use it as valid data.  Further, Xastir
is going through the trouble of COMPUTING the gust from the history of
wind speeds, and this means we're clobbering what it computed every
time we get a data packet from the weather station.

This commit conditionalizes the setting of wx_gust (and the premptive
memsetting of that field) on whether we got valid "g" data.  If it
finds "..." after the g, we just don't do anything.

This makes Xastir now report meaningful gust data at all times,
and it doesn't get wiped out every time there's an incoming report
from the attached weather station.

This should have no effect on stations connected to real Davis devices
that report gust (because we only ignore it if the gust data is
"...").  And it should be OK for wxnowsrv.pl users as well, because
that script just dutifully turns the wxnow data into a Davis-like
string.  If that string contains valid gust data, we're good, and if
it has "g..." then we ignore that bit.