Remove install-xastir script
install-xastir is a script designed "for those folks challenged by the
Unix way of doing things."  All it does is a git clone of Xastir and a
call of the "update-xastir" script.

It was created in 2004 in commit 02310b5, and is past its sell-by
date.

Since one has to clone Xastir to get the script, having a script that
clones Xastir and then runs a different script is pointless, and the
name of the script is too tempting --- a user who is rushing to try to
install the code for the first time without reading any docs might
actually run it.  That happened today.