Fix autoconf probe for motif library
The code for probing Motif tried to link a small test program that
uses XmGetDestination.  This was done with no prologue including any
Motif headers, nor with all the correct arguments for
XmGetDestination.

If a user is compiling Xastir using any CFLAG that includes "-Werror",
or if the compiler is excessively strict, this fails.

This commit adds the prologue needed to get the correct prototype for
XmGetDestination, then calls it with the correct prototype in the test
program.

This eliminates the warning on my system and should eliminate the
failure on systems where -Werror has somehow been forced into the
CFLAGS or otherwise enabled.  It seems to be that way for at least one
Mac user after a recent upgrade.