Update proj probe to look for modern function
Recent versions of libproj deprecated the decades-old PROJ.4 API in
favor of a much more flexible and modern version.

Xastir has always looked for one of the functions (pj_init) in the original API
in its configure probe.  Since this function will be going away in
proj 7 we can't just keep doing that.

Now, configure probes for the function proj_create_crs_to_crs
instead.  If it finds it, we have a newer libproj and we use it.  If
it doesn't, it tries again with pj_init -- and if it finds *THAT* then
we have an older libproj and use it.

Xastir doesn't even use libproj.  Our proj probe is very ancient and
is there because in the old days no package systems had geotiff and
everyone who wanted geotiff support had to build it from source
themselves.  That generally wound up creating static libraries, and
therefore one had to specify all the dependencies explicitly.  Geotiff
requires proj, so to get geotiff working from a static-built library
you also have to link proj with Xastir.

These days every package management system has geotiff and it's a
shared library, so proj is pulled in automatically at run time.  This
lingering probe is there only to support hardcore folks who build
everything themselves from source.

Now it will keep working even when proj 7 starts  showing up in
package systems.

Closes #173