Remove update:true from OS X config
A while back, in commit 38c0313 on 21 Feb 2020, we added an option to
the Travis config for OS X to force a homebrew update.  This fixed a
temporary problem they were having there where homebrew was outdated,
but bloats the build time dramatically (like to over 20 minutes).

I am removing it in the hopes that it is no longer needed.