Revert "Remove update:true from OS X config"
This reverts commit 9bb89b37d0ec51fb182d5a37a281e969e8ead357.

Unfortunately, the "update: true" thing I commented out in the
reverted commit is still necessary, Travis still has a broken VM for
OS X with an outdated homebrew.

This means our OS X builds take an enormous time, because not only
does it have to build Xastir, it has to update the entire homebrew
system.