Update OS X image choice
The Travis default osx_image is "xcode9.4" which is truly ancient, and
we weren't using it.  We used xcode10.2.  This is old enough that
homebrew requires constant update and means our build takes forever.

I'm trying to update the osx_image choice to the very latest to see if
it fixes our build time and still actually works.

I am turning off the update:true flag at the same time.