Finish removing buggy, memory-unsafe debugging code
In commit 5979890 from 1 April 2003, some debugging fprintfs were
turned into sprintfs and strncats into fixed-length buffers, followed
by a single fprintf at the end (the fprintf was then commented out later
that same day in commit b408025).  The buffers are used for nothing
except this output.

The sprintfs into the fixed buffer do no checking to make sure they
don't overrun the buffer (though the strncats do).

In #174, it was reported that one of these sprintfs was causing a
segfault when the code was compiled with certain memory checking
turned on.  Pull request #175 removed one of the problematic sprintfs,
this commit removes the remainder of them and the declaration of the
buffers.