Remove debugging output to STDERR
In commits  a7c0ad0 Curt uncommented a debugging fprintf and in
d5d4a2b he added another.

Neither of these are important for a user to see, and in fact look
like an error to them.  They should have been removed when the
debugging was done.

I am removing them now.