Update documentation to reflect new release process
I've proposed doing away with the "prebootstrapped" release tarball idea,
which simplifies how we make point releases, but means that *all*
users now need to bootstrap, no matter how they get the code.

This commit changes the README.developers.md file to explain how the
new process would work and the INSTALL document to remove the "if you
got from a tarball, you're ready to go, but if you're using git, you
need to bootstrap" condition, and replace it with "you need to run
bootstrap".