Update README.Getting-Started
I am updating README.Getting-Started to show that autoconf and
automake are now mandatory packages for all users, because all
users (tarball and git) must now use "bootstrap.sh" before trying to
configure xastir.

Furthermore, this file was last updated (it appears) when we first
moved to git around release 2.1.0, and was not updated when the names
of tarballs was changed (right before 2.1.0!).  It was only very
briefly that our tarball names had two "Xastir"s in them, and we never
did a proper release with them that way.