Start using "git describe" in Help dialog version
Prior to our move to git, Xastir used to put the date it was compiled
into the Help dialog.  We stopped doing that, partly because it was
not that informative, but also because it made the build
"non-reproducible" (two consecutive builds of Xastir from the same
source code produced binaries that were not byte-for-byte identical).

For a while now, we've been inserting the git SHA-1 from which the code
was built into the Help dialog.  This only happens if the Xastir
source tree has a ".git" directory --- so it will not happen if Xastir
is built from a tarball, or if someone has unpacked Xastir sources
into their own git repo (as is done by some Linux packaging systems).

The purpose of putting the git sha in there was so that users of
Xastir from the development git repo could know how current their
build actually is just by looking at the Help->About dialog and seeing
if it's the same as the one in their git tree.

Until now, Xastir's release tags have all been on dead-end branches,
or were old CVS release tags.  Thus, the most information we could get
out of git was the SHA-1 of the current commit being built.

Now that we've changed our release process so that release tags are
reachable from the head of master, "git describe" becomes a more
informative way of identifying the code.

I am changing XastirDateStamp.sh to populate the gitstring variable
with the output of "git describe --dirty" instead of the old git log
command it used.  This will now populate the Help dialog with a string
like "Release-2.1.8-1-g8de3f8a1-dirty", which means "1 commit past
release 2.1.8, git SHA g8de3f8a1, with uncommitted code present".

Now you can not only tell what the last commit was, but whether
you're running *that* code or some modified code, and how many commits
have happened since the last release.