Fix issue 121 for real
Issue #121 asked us to support more GPS RMC and GGA strings than just
those that start with '$Gx" (where x is P, N, A, L, or B).  The user
specifically wanted us to recognize "$II" strings.

Pull request #122 got this almost there by taking out the probe for
the initial "$G" and looking only for "$", but left in place the check
for the P,N, A, L or B in the third character.

This commit removes that final check, which should actually fix #121
instead of just leaving it half fixed.

Sadly, the original requester never got back to us to tell us that the
fix didn't work, and so it went unnoticed until today when I was
reviewing the old PR and this glaring error jumped out at me.

Xastir#121
Xastir#122
Xastir#53
Xastir#58