Remove unneeded configure probes
Xastir's configure.ac was probing for argz.h and several "__argz_"
functions, as well as "dcgettext".

Probing for dcgettext was pointless, as all use of gettext and libintl
was removed back in 2019 in commits 04b3071 and 6c81a23.

The argz functions were part of the vestigial use of libintl.  All use
of that stuff was removed in 04b3071, too.

Probing for this stuff should have been removed at the same time, but
got overlooked.