Remove derived language generation file from boostrap
In commit 21ebe38, Curt added generation of derived language
files (mostly joke languages like Pig Latin and Pirate English) to
bootstrap.sh, and slightly tweaked that technique the next day in
commit 8d52f59.

Two years later, in commits aae6102 and 67ed45c, Gerald Stueve modified
config/Makefile.am to build these language files at build time instead
of at bootstrap time, which is a better approach.  Unfortunately, this
was done with little fanfare *and* Gerald did not remove the redundant
regeneration in bootstrap.sh.

The presence of non-autoconf stuff in bootstrap.sh was weird, so it
was right to move it elsewhere.  Now, bootstrap.sh is not much
different (or at all different) from using "autoreconf", the newer,
less kludgy way of "bootstrapping" that replaced the shell script in
many projects.

This came up during an attempt to get FreeBSD's Xastir port updated to
2.1.8.  The new need to run bootstrap.sh even when using tarballs from
github complicated the port and made it a mess.  But using autoreconf
is easy.

Fortunately, the generation of language-PirateEnglish.sys and its ilk
was already being taken care of at "make" time, so the port maintainer
chose to skip the mess of using bootstrap and just set up the port to
use autoreconf.

This commit removes the language generation from bootstrap, which
should have been done at the same time that capability was put into
config/Makefile.am.

https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=259669