Add autoreconf information
Bootstrap.sh is an ancient technique that used to be recommended in
documentation about "autoconfiscating" projects for portability.  It
was meant to run all the bits of the autotools (aclocal, autoheader,
autoconf, automake...) to create the configure script.

Some time much later than Xastir's bootstrap.sh was created, autotools
began to include a program called "autoreconf" that does basically the
same thing.  As it happens, this approach *does* work with Xastir,
although for a period of about two years it would not have.

Our bootstrap.sh got a few non-autoconf steps added to it in 2008,
which were then quietly superseded by changes to config/Makefile.am
that *should* have led to those steps being removed from our
bootstrap, but didn't.

Our bootstrap.sh *now* has ONLY autoconf stuff in it, and so is
now basically identical to running "autoreconf -i".

This commit adds a little information about using autoreconf instead
to several of the document files (README.developers.md, README.GIT,
INSTALL, and FAQ).