Beef up documentation of xastir_udp_client
The documentation of xastir_udp_client never explicitly stated what it
expected as "APRS Packet Goes Here".  What it means is that a COMPLETE
packet, including from and to calls, goes here.

This vagueness has led to one Xastir user thinking that the utility
could be used by specifying only the APRS payload for the packet (in
his case, the payload was a perfectly valid weather station payload,
produced by the WeeWX aprs extension) without adding the necessary
packet header (apparently other tools for injecting packets into APRS
add that for you, such as the Linux "beacon" utility or aprx's
"beacon" feature).

This commit simply adds notes to the existing barebones "NOTES"
section of the xastir_udp_client man page stating that the final
argument given to the utility must be a complete packet, including the
"FROM>TO:" header.

Xastir#189