Add commentary about object/item adoption
If a user uses xastir_udp_client to inject a packet that contains an
object or item payload, and the callsign/ssid of the packet matches
Xastir's, then Xastir will adopt the object or item as its own and
retransmit it the same way it would have had the object been created
through its own interface.

This can be either extremely useful or really surprising depending on
whether one knew it would happen or not.  Document it.