configure.ac: use AC_PROG_EGREP
"egrep" is an obsolete alias for grep -E and newer greps will warn on usage
of egrep, so let's just swap it out where the newer "grep -E" is available.

It looks like "grep -E" was added to POSIX.. 2001? at the earliest, but it's
easy to play it safe here, so why not?

Signed-off-by: Sam James <sam@gentoo.org>