Remove egrep usage from configure.ac
In commit fd3a1c6 in 2005, a bit of code was added to configure.ac
that ran an "egrep" script to extract certain lines from config.log
into summary.log.  Summary.log was previously only used by the
"REGRESSION_TESTS" script to dump out the list of libraries compiled
into a binary (the same list used in the Help->About box).

This was theoretically to make diagnosing user problems easier,
because it extracted relevant "checking..." and "result" lines and
threw out all the rest of config.log.  But in the 17 years since it
was committed, I have never actually seen anyone say "send me your
summary.log" file when diagnosing a user problem, they ask for the
whole config.log anyway.  And this is how it should be.

In pull request #194, there was a request to replace the antiquated
"egrep" usage with "grep -E" that has been part of POSIX since 2001.
On most systems nowadays, egrep simply invokes grep -E anyway.  More
recent versions of grep have started emitting deprecation warnings
when invoked as egrep, so it is not desirable to hard-code "egrep"
calls into configure.ac.

While that pull request did indeed make changes to configure.log that
was portable (it would fall back to egrep if grep -E didn't work), it
seems more prudent just to remove the egrep use altogether, since it
serves no purpose.
Closes #194