Update get-NWSdata and associated dbfawk
NWS has changed its data files again, and has made a change to the
field names in the "w" file.  The only change to the field name is
that they're all upper case now, whereas previous versions were mixed
case.

This commit updates get-NWSdata to make it download the new shapefiles
and adds a new dbfawk file to make it work with the new "w" shapefile.

Closes #196