Clean up configure test for no-keep-memory
In #204, we accepted a pull request that fixed a C99 incompatibility
issue in configure.ac.

The pull request was fine and got the real problem addressed, but was
the minimal change necessary to do the job.

This commit removed the now-unnecessary include of sys/types.h
altogether, and adds configure noise to say what it is doing at that
point.

I strongly believe this entire test should be removed, as it forces
the link phase of all builds of Xastir to be optimized for memory
usage instead of speed.  It was put there in 2004 so that Cygwin users
wouldn't have to do it themselves, but as a result forced it into the
linker flags for everyone.  Things like this should not be forced,
they should be provided to configure in the command line only when
needed. Further, features of configure like this are maintenance
problems and could become outdated very quickly (like, when memory
becomes so cheap that even Cygwin doesn't need this kind of thing).