Remove some very old commented-out code
Two bits of commented-out code are removed here.  One is a block of
different settings of converse string that were made obsolete long ago
by a new configuration option in 2011, and another is a single line of
code that was commented out in 2002.

Updated some verbose comments that contained history of the block that
is of no import, and can be explained in git history anyway.