Remove another commented-out code block
In my last commit I removed a line of code that was commented out in
2002, but it was an if statement with a brace.  This commit removes
the commented-out "} else {...}" that went with it.  I also removed
the part of the comment that said a bit of code had been commented
out, but left the explanation of what the code that *wasn't* commented
out was doing.