Simplify conditionals regarding shape/pcre
Back when we had a hard-coded alternative to DBFAWK in map_shp.c, we
had to distinguish between a compilation where shapelib was found and
pcre wasn't, and when both were found.

In commit 3a9b6e17 the hard-coded rendering of shapefiles was removed,
and now configure refuses to compile in shapefile support if pcre is
not found.

It is now therefore unnecessary to have ifdefs that look for
"HAVE_LIBPCRE" because if "HAVE_LIBSHP" is defined then
configure *also* found a suitable PCRE.  It is enough to ifdef based
on HAVE_LIBSHP alone.

This commit removes all the ifdefs like "#if defined(HAVE_LIBSHP) &&
defined(HAVE_LIBPCRE)" or "#ifdef HAVE_LIBPCRE", replacing both with
"#ifdef HAVE_LIBSHP"

This commit is groundwork for really addressing issue #203