Isolate old PCRE API calls and data
When I add PCRE2 support, there are lines of Xastir's awk.c and awk.h
that will need to be completely overridden by new API calls and
declarations.

This commit adds configure.ac defines to flag that a legacy PCRE is
being used, and puts all calls to that legacy interface into an ifdef.

I have confirmed that doing this has absolutely no effect on a build
that already worked with legacy PCRE, and the "testdbfawk" program
works just fine to produce exactly the same output it always did on
input it always handled before.  That is, it properly parses the
DBFAWK patterns and actions, and produces valid assignment of Xastir's
various map rendering variables.  Xastir renders shapefiles
correctly (or rather, as correctly as it ever did).

More groundwork for issue #203