Add PCRE2 support
This commit changes configure so that it first looks for pcre2, and if
it finds it it uses it.  Only if pcre2 cannot be found (either because
the header or library doesn't work) then the legacy pcre is searched.

All code that uses PCRE has been conditionalized to do the right thing
no matter which version is linked to.

I have tested it in both legacy and current versions.

Fixes issue #203