Remove AC_HEADER_STDC and clean up AC_CHECK_HEADERS calls
AC_HEADER_STDC is obsolete, and all new versions of autoconf scream
loudly at us for  using it.  All it does is check whether a handful of
standard header files are present --- and on all modern
standard-compliant compilers they are.

Further, AC_CHECK_HEADERS will, for each listed header file, create
and compile a small test program to check whether the header (say
"foo.h") exists or not, and if it DOES exist define a macro
HAVE_FOO_H.  It does not cause configure to exit with an error if the
header file is not found.  The intent is that programs will ifdef
includes of that header using the provided macro, and find a
workaround if the macro isn't defined.

In many, many cases, headers probed by long lists in AC_CHECK_HEADERS
were either not even being used in Xastir at all, or were being
included unconditionally, or worse, were included unconditionally in
some places and conditionally in others.  The effect of this would be
that Xastir builds would crash in the compile phase if the header
didn't exist.  Since we have not had complaints about this happening
in decades, I have to assume that all of the systems that current
Xastir users are actually using anymore all have those files and we
don't need to probe for them.

I have left in place any such header checks where we actually *do*
consistently check the macro and do correct things when it isn't
found.

I have removed all the includes of "strings.h" that were put there
only for Solaris users, because by removing AC_HEADER_STDC we're
pretty much saying "Xastir will no longer compile on any system that
doesn't comply with C89 or C99 standards and doesn't provide headers
mandated by those standards."  If you're still using a Solaris system
you pulled out of a surplus store 20 years ago, you'll have to stick
with versions of Xastir prior to this commit.

Partially  addresses #202