Remove AC_TYPE_SIGNAL macro usage
AC_TYPE_SIGNAL is obsolete, as since C89 the standard return type for
"signal" is supposed to be "void".

AC_TYPE_SIGNAL sets RETSIGTYPE to "void" if signal.h defines "signal"
to be a "void" and to int otherwise.

Only one file in Xastir actually used RETSIGTYPE at all or ifdefed
anything related to it.  In fact, in almost every single case where
"signal" was used it was *assumed* to be returning a void and was cast
to that anyway, unconditionally, ignoring the setting of RETSIGTYPE.

Worse, in a number of places the existance of RETSIGTYPE was
apparently a problem somewhere and it was being explicitly undeffed.
This has been the case since before Xastir source code was ever put
into CVS on sourceforge in 2002.

This commit removes the AC_TYPE_SIGNAL macro from configure, removes
any conditional use of RETSIGTYPE, all the undefs of RETSIGTYPE, and
in the ONE place in the entire code where RETSIGTYPE was being
used correctly (forked_getaddrinfo.c), we just put in a "#define
RETSIGTYPE void".

Partially addresses #202