Remove AC_HEADER_TIME usage
The AC_HEADER_TIME macro checks if the compiler/system allows both
sys/time.h and time.h to be included without conflict.  It has been
standard for this to be possible if both exist for quite some time,
and using this macro should be unnecessary.  The AC_HEADER_TIME macro
is considered obsolete by autoconf, which complains loudly if you use
it.

To  remove this macro call I had to edit every file that made use of
the TIME_WITH_SYS_TIME macro that is set by AC_HEADER_TIME.

In all cases, this block:
 #if HAVE_SYS_TIME_H
    #include <sys/time.h>
    #include <time.h>
 #else
   #if HAVE_SYS_TIME_H
     #include <sys/time.h>
   #else
     #include <time.h>
   #endif
 #endif

has been replaced with:

   #if HAVE_SYS_TIME_H
     #include <sys/time.h>
   #endif
   #include <time.h>

That is, we always include sys/time.h if it exists, and always include
time.h.

In exactly ONE place (gps.c) there was a claim that one needed to
define "__USE_XOPEN" for "some" linux systems before including
sys/time.h in order for strptime to work, but this is also very
obsolete, as strptime should just be available.

Worse, db.c uses strptime which it gets only because alert.h includes
sys/time.h and time.h *without* this macro being defined --- so
clearly it was already inconsistently applied.  So since we've not
been getting complaints all over the place about strptime not being
found, the "__USE_XOPEN" thing is clearly not ever a problem for our
users and is not needed anymore.  I have removed it.

Last changes addressing #202