Remove workarounds for ancient versions of ImageMagick
In olden times, the ImageMagick API was quite unstable and kept
changing out from underneath us.  Because it had not yet made it into
package management systems, most users of Xastir that needed
ImageMagick at the time *also* needed to compile it themselves from
source.  As a result, many of them would not want to recompile the
very latest version of ImageMagick every time an API got broken, but
neither would new users want to have to stick to an old version.
Xastir code was therefore peppered with all manner of ifdefs to handle
every old version of ImageMagick one might have installed.

The first of these ifdefs was already present in 2002 when Xastir was
first imported into CVS.  New ifdefs were added in 2002, 2003, and a
final set in 2018.  There have been no more new versions of
ImageMagick with broken APIs since then that we can use.

ImageMagick 6 is already ancient, and Xastir can't work at all with
ImageMagick 7.

Maintaining ifdefs to support 20 year old versions of ImageMagick just
bloats the code and makes it unreadable.  It isn't even necessary to
continue to support those that were already old in 2018.

I have removed all the ifdefs that deal with any version of
ImageMagick older than the one that reports its MagickLibVersion as
0x0669.  There have been no API breakages since then, and we can
safely assume now that all users of ImageMagick 6 have a version at
least as new as that.

ImageMagick support in Xastir is not long for this world, since
ImageMagick's focus seems to be pushing harder into HDRI, which is
completely incompatible with our image usage and would require a
complete rewrite of the Magick code.  GraphicsMagick, however,
continues to pursue stability of API rather than cutting edge image
processing, and even if ImageMagick6 disappears from every single
Linux package management system, we can still work so long as
GraphicsMagick exists.

GraphicsMagick is already our recommended tool and we always use it if
it is found and the user hasn't explicitly requested ImageMagick.
We don't even look for ImageMagick unless the user has explicitly
disabled graphicsmagick *or* graphicsmagick isn't found.

There are still a few ifdefs to deal with ONE GraphicsMagick API
breakage, and someday we should probably remove that, too.  But issue
 #25 was strictly about removing "old ImageMagick API breakage
workarounds", so that's all I'm doing here.

Addresses #25