Issue 199 -- Move code to fix locate callback
The setup of the callback when an icon is clicked in the station
list dialogs uses the callsign from the station name column to
locate the desired station. The name was update after the callback
was set, so in a changing environment the callback would display
the location of the station that previously occupied that row in
the list.

Move the updating of the name to be before setting up the icon.