Tidy up some README documents
README.developers.md, used by developers at release time, needed a
tiny bit of work.

README.Getting-Started, which was intended as a "quick start" document
but really has a lot more in it than that, had some obsolete contents
that I have cleaned out.  For example, we no longer need to tell users
to install CVS and RCS, we need them to know that GraphicsMagick is
preferred to ImageMagick, and that if they MUST use ImageMagick it
must be ImageMagick6.  Finally, they need to know that PCRE2 is now
preferred over PCRE because some package management systems are
starting to drop PCRE (which has been past end of life since 2015).

Doing this in preparation to having to do a release 2.2.0 next week.