Mess around with the "INSTALL" file a bit
INSTALL is what we tell users to look at in the README.md file to get
Xastir built.

Honestly, I don't think this document is fit for the average new user,
as it looks an awful like a compilation of random thoughts taken out
of a decade or so of miscellaneous user support emails from the
mailing list, bouncing between a general "here's how to build" to an
in-depth description of the interplay between various libraries and
the features they enable, followed by a set of "First install"
instructions that include telling users to build a bunch of dependent
libraries from source, with all manner of side excursions into
optional features.  It is very much a creature of the early 2000s,
when almost every library we use was a niche item that appeared in no
package managers.  It could use a complete rewrite.

I have not completely rewritten it.

I have added some text at the top to point potential users to the
Wiki's "Installation Notes" for system-specific guidance, almost all
of which may be more helpful than either the general, vague guidance
in this file or the detailed "drink from the firehose" stuff that
follows it.

I also removed several references to configure options that no longer
exist, rearranged the lists of libraries so one can tell the
difference between what is absolutely needed before proceeding with
the general guidance at the beginning and what is optional.

And finally, I put a dividing line between the general guidance and
the stuff I feel is too much information for a first-time install, or
which may very well be completely outdated (like, telling folks where
to download libgeotiff source and how to compile proj libraries).