Try to create a simplified "getting started" document
README.Getting-Started and INSTALL are old documentation files that do
not read to me like a quick start "here's how to build for the first
time" document, and I really want to simplify them.

I have created a new INSTALL.md file and modified the README.md so
that I think they may be a bit more readable and accessible.