Fix altitude-in-comment with PHG
Xastir was improperly putting "A=xxxxxx/" into the comment field of
position reports with PHG.  This is only done for fixed stations with
GPS attached (the only ones that have altitude to report).  The reason
it doesn't get done correctly is that there was a logic error in the
code that was intended to insert a slash before the altitude in just
the fixed-station+PHG cases.

It has been wrong since at least 2002.

The problem is that the spec requires "/A=xxxxxx" for the altitude to
be recognized.

In other parts of the code, where we're adding altitude to course and
speed, we were unnecessarily putting a "/" at the end of course and
speed, then conditionally adding "/" in between that and the altitude
string depending on whether it had already been added or not.  This is
unnecessarily complicated, as the "/" should ALWAYS be attached to the
altitude string, and in fact the spec does not even require that
course and speed be followed by a slash.

The code that tried to insert the slash for PHG was broken and was not
inserting one.

However, rather than fix the broken logic, I have chosen to simplify
the code and keep it spec compliant.

This commit puts the slash at the beginning of the altitude string
unconditionally, removes the unconditional slash from the end of
course and speed, and removes all the attempted logic that was
intended to insert the slash in between course/speed or PHG and the
altitude.

I have confirmed that in every case where altitude was previously
correct that it remains correct, and that in the case of PHG it is now
correct where it had previously been wrong.

In doing so, I have also ripped out three identical copies of the code
to format a fixed station report and refactored it into a function
that is called three times.  Cut-and-paste code is bad form as it is
harder to maintain three identical copies of code than one reusable
element.

Addresses #198