Update a comment
map_shp.c (which I'm currently studying with an eye to tidying it up
and fixing issue #154), had a comment inserted during the original
creation of DBFAWK in 2003 (commit 62a8be7).  This comment referred to
"all this WX junk following" that was intended to discuss how DBFAWK
differed from the non-dbfawk code that was still there and enabled or
disabled by ifdef.

That code was deleted a long time ago in commit 3a9b6e17, but the
comment remained.

Now, instead of referring to code that no longer exists, it describes
what the code that still exists does.